+++ IEntity.go
@@ -39,8 +39,8 @@ type IEntityStruct interface {
 	GetPkSequence() string
 
 	// GetDefaultValue 获取列的默认值Map,用于Insert和Update,UpdateNotZeroValue Struct对象,返回map的key是Struct属性名,value是默认值,value可以是nil,不能是类型的默认值,比如int类型设置默认值为0
-	// 重点说明:UpdateNotZeroValue方法也会更新GetDefaultValue设定的默认值.特殊情况不想更新,使用BindContextDefaultValue废弃掉GetDefaultValue方法.全局情况可以OverrideFunc复写UpdateNotZeroValue.
-	// GetDefaultValue To get the default value of the Map, for the Insert and Update,UpdateNotZeroValue Struct objects,  The key that returns map is the Struct property name, value is the default value, and value can be nil.
+	// 重点说明:UpdateNotZeroValue不会取值DefaultValue
+	// GetDefaultValue To get the default value of the Map, for the Insert and Update Struct objects,  The key that returns map is the Struct property name, value is the default value, and value can be nil.
 	GetDefaultValue() map[string]interface{}
 }
 
@@ -105,9 +105,9 @@ func (entity *EntityStruct) GetPkSequence() string {
 
 // GetDefaultValue 获取列的默认值Map,用于Insert和Update,UpdateNotZeroValue Struct对象,返回map的key是Struct属性名,value是默认值,value可以是nil.不能是类型的默认值,比如int类型设置默认值为0
 // BindContextDefaultValue 优先级高于 GetDefaultValue
-// 重点说明:UpdateNotZeroValue方法也会更新GetDefaultValue设定的默认值.特殊情况不想更新,使用BindContextDefaultValue废弃掉GetDefaultValue方法.全局情况可以OverrideFunc复写UpdateNotZeroValue.
+// 重点说明:UpdateNotZeroValue不会取值DefaultValue
 
-// GetDefaultValue To get the default value of the Map, for the Insert and Update,UpdateNotZeroValue Struct objects,  The key that returns map is the Struct property name, value is the default value, and value can be nil.It cannot be the default value of the type, for example, the default value of the int type is set to 0
+// GetDefaultValue To get the default value of the Map, for the Insert and Update Struct objects,  The key that returns map is the Struct property name, value is the default value, and value can be nil.It cannot be the default value of the type, for example, the default value of the int type is set to 0
 // BindContextDefaultValue takes precedence over GetDefaultValue
 func (entity *EntityStruct) GetDefaultValue() map[string]interface{} {
 	return nil