+++ CHANGELOG.md
@@ -3,6 +3,7 @@ v1.7.0
  - IsBindDBConnection 检查ctx是否已经绑定数据库连接
  - Finder增加GetValues方法,返回Finder封装的values值
  - zormErrorSQLValues 使用json格式输出
+ - BindContextOnlyUpdateCols 指定仅更新的数据库字段,只对Update方法有效.cols是数据库列名切片
  - 完善文档,注释
 
 v1.6.9