+++ CHANGELOG.md
@@ -5,7 +5,7 @@ v1.7.0
  - zormErrorSQLValues 使用json格式输出
  - BindContextOnlyUpdateCols 指定仅更新的数据库字段,只对Update方法有效.cols是数据库列名切片
  - BindContextMustUpdateCols 指定必须更新的数据库字段,只对UpdateNotZeroValue方法有效.cols是数据库列名切片
- - BindContextDefaultValue 替代 BindContextMustUpdate函数, 优先级高于 GetDefaultValueMap
+ - GetDefaultValueMap更名为GetDefaultValue. BindContextMustUpdate更名为BindContextDefaultValue, 优先级高于 GetDefaultValue
  - FuncWrapFieldTagName函数增加*reflect.StructField参数
  - 完善文档,注释
 
@@ -15,7 +15,7 @@ v1.6.9
 
 v1.6.8
  - 自动生成CountFinder.InjectionCheck和查询Finder保持一致
- - IEntityStruct接口增加GetDefaultValueMap函数,设置Struct属性默认值
+ - IEntityStruct接口增加GetDefaultValue函数,设置Struct属性默认值
  - 完善文档,注释
 
 v1.6.7