+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+v1.7.1
+ - UpdateNotZeroValue不会取值DefaultValue
+ - 完善文档,注释
+
 v1.7.0
  - Finder和EntityMap增加序列化json方法,适配兼容wasm环境.
  - IsBindDBConnection 检查ctx是否已经绑定数据库连接